==> 01_eda.ipynb <==
# Imports

import pandas as pd

# Load data

df_tr_1 = pd.read_excel('data/raw/Banking transactions.xlsx', sheet_name=0)
df_tr_1

df_tr_2 = pd.read_excel('data/raw/Banking transactions.xlsx', sheet_name=1)
df_tr_2

df_tr_3 = pd.read_excel('data/raw/Banking transactions.xlsx', sheet_name=2)
df_tr_3

# Health checks

Since we have no date, I guess it's natural to think that duplicates is okay

display(
    df_tr_1[df_tr_1.isna().sum(axis=1) != 0]
    , df_tr_2[df_tr_2.isna().sum(axis=1) != 0]
    , df_tr_3[df_tr_3.isna().sum(axis=1) != 0]
)

pd.concat([df_tr_1, df_tr_2, df_tr_3])

display(
    df_tr_1.columns
    , df_tr_2.columns
    , df_tr_3.columns
)

df_tr_3.rename(columns={'Category ': 'Category'}, inplace=True)
df_tr_3.columns

df = pd.concat([df_tr_1, df_tr_2, df_tr_3])
df

# EDA

df[df.duplicated()]

Since we do not have full set of primary keys - it seems okay to have duplicates

df.isna().sum()

df.describe()

df['Category'].value_counts()
