==> 03_split_test.ipynb <==
# Imports

import pandas as pd
from sklearn.model_selection import train_test_split

# Load data

df = pd.read_csv('data/processed/concat.csv')
df

# Split test

df_train, df_test = train_test_split(df, test_size=0.1, random_state=10)
display(
    df_train.shape, df_test.shape
    , df_train.head(), df_test.head()
)

# Save

df_train.to_csv('data/processed/train.csv', index=False)

df_test.to_csv('data/processed/test.csv', index=False)
