==> 04_data_prep.ipynb <==
# Imports

import pandas as pd
import pickle
from sklearn.feature_extraction.text import TfidfVectorizer

# Load data

df = pd.read_csv('data/processed/train.csv')
df.head()

# Prepare

### 1) Drop duplicates

As it was decided in 01_eda - we won't drop duplicates cause they are most likely different transactions

### 2) Fillna

fillna_values = {}

for col in ['Description', 'Amount']:
    fillna_values[col] = df[col].mode()[0]

fillna_values

for col in fillna_values.keys():
    df.loc[df[col].isna(), col] = fillna_values[col]

df

### 3) Encode categoricals

class CategoricalEncoder:
    def __init__(self, df, columns=None):
        self.dtypes = df.dtypes.apply(str).to_dict()
        if columns is None:
            self.columns = [k for k in self.dtypes.keys() if self.dtypes[k] == 'object']
        else: 
            self.columns = columns
        self.mapping = self.__get_mapping(df, self.columns)
        
        
    def __get_mapping(self, df, cols) -> dict:
        mapping = {}
        for c in cols:
            unique_cats = df[c].unique()
            # assign to each unique value its number
            mapping[c] = dict(zip(unique_cats, range(len(unique_cats))))
        return mapping
    
    def encode(self, df, col) -> None:
        assert col in self.columns, 'Unknown column. Cannot process further'
        # values not in mapper
        df.loc[~df[col].isin(self.mapping[col].keys()), col] = -1
        # values in mapper
        df[col].replace(to_replace=self.mapping[col], inplace=True)
    
    def decode(self, df, col) -> None:
        assert col in self.columns, 'Unknown column. Cannot process further'
        inv_mapping = {v: k for k, v in self.mapping[col].values()}
        df.loc[~df[col].isin(inv_mapping.keys()), col] = None

enc = CategoricalEncoder(df, ['Category', 'Currency'])

for c in enc.columns:
    enc.encode(df, c)
    
df.head()

unique_cats = df['Category'].unique()

dict(zip(unique_cats, range(len(unique_cats)))).values()

### Description col

tf_idf = TfidfVectorizer(max_features=50)

tf_idf_matr = tf_idf.fit_transform(df['Description'])


df_tf_idf = pd.DataFrame(tf_idf_matr.toarray(), columns=tf_idf.get_feature_names_out())
df_tf_idf

df_concat = pd.concat([df, df_tf_idf], axis=1)

df_concat

# Save

df_concat.to_csv('data/ready/train.csv', index=False)

with open('models/tf_idf.pickle', 'wb') as handle:
    pickle.dump(tf_idf, handle)

with open('models/fillna.pickle', 'wb') as handle:
    pickle.dump(fillna_values, handle)

with open('models/enc.pickle', 'wb') as handle:
    pickle.dump(enc, handle)
