==> 05_rf.ipynb <==
# Imports

import pandas as pd
from sklearn.model_selection import train_test_split
from sklearn.ensemble import RandomForestClassifier

from sklearn.metrics import accuracy_score, confusion_matrix

import pickle
from sklearn.feature_extraction.text import TfidfVectorizer

class CategoricalEncoder:
    def __init__(self, df, columns=None):
        self.dtypes = df.dtypes.apply(str).to_dict()
        if columns is None:
            self.columns = [k for k in self.dtypes.keys() if self.dtypes[k] == 'object']
        else: 
            self.columns = columns
        self.mapping = self.__get_mapping(df, self.columns)
        
        
    def __get_mapping(self, df, cols) -> dict:
        mapping = {}
        for c in cols:
            unique_cats = df[c].unique()
            # assign to each unique value its number
            mapping[c] = dict(zip(unique_cats, range(len(unique_cats))))
        return mapping
    
    def encode(self, df, col) -> None:
        assert col in self.columns, 'Unknown column. Cannot process further'
        # values not in mapper
        df.loc[~df[col].isin(self.mapping[col].keys()), col] = -1
        # values in mapper
        df[col].replace(to_replace=self.mapping[col], inplace=True)
    
    def decode(self, df, col) -> None:
        assert col in self.columns, 'Unknown column. Cannot process further'
        inv_mapping = {v: k for k, v in self.mapping[col].values()}
        df.loc[~df[col].isin(inv_mapping.keys()), col] = None

# Load data

df = pd.read_csv('data/ready/train.csv')
df

df.drop(columns=['Description'], inplace=True)

# Train-val split

df_train, df_val = train_test_split(df, test_size=0.1, random_state=10)
display(
    df_train.shape, df_val.shape
    , df_train.head(), df_val.head()
)

# Apply model

df_train.drop(columns='Category')

rf = RandomForestClassifier(n_jobs=-1, random_state=10)

rf.fit(df_train.drop(columns='Category'), df_train['Category'])

y_train_pred = rf.predict(df_train.drop(columns='Category'))
y_val_pred = rf.predict(df_val.drop(columns='Category'))

print('Train accuracy score: '
      , accuracy_score(df_train['Category'], y_train_pred)
)
print('Val accuracy score: '
      , accuracy_score(df_val['Category'], y_val_pred)
)

rf = RandomForestClassifier(n_jobs=-1, random_state=10, min_samples_leaf=3)

rf.fit(df_train.drop(columns='Category'), df_train['Category'])

y_train_pred = rf.predict(df_train.drop(columns='Category'))
y_val_pred = rf.predict(df_val.drop(columns='Category'))

print('Train accuracy score: '
      , accuracy_score(df_train['Category'], y_train_pred)
)
print('Val accuracy score: '
      , accuracy_score(df_val['Category'], y_val_pred)
)

rf = RandomForestClassifier(n_jobs=-1, random_state=10, min_samples_leaf=10)

rf.fit(df_train.drop(columns='Category'), df_train['Category'])

y_train_pred = rf.predict(df_train.drop(columns='Category'))
y_val_pred = rf.predict(df_val.drop(columns='Category'))

print('Train accuracy score: '
      , accuracy_score(df_train['Category'], y_train_pred)
)
print('Val accuracy score: '
      , accuracy_score(df_val['Category'], y_val_pred)
)

rf = RandomForestClassifier(n_jobs=-1, random_state=10, min_samples_leaf=1, max_features=0.5)

rf.fit(df_train.drop(columns='Category'), df_train['Category'])

y_train_pred = rf.predict(df_train.drop(columns='Category'))
y_val_pred = rf.predict(df_val.drop(columns='Category'))

print('Train accuracy score: '
      , accuracy_score(df_train['Category'], y_train_pred)
)
print('Val accuracy score: '
      , accuracy_score(df_val['Category'], y_val_pred)
)

rf = RandomForestClassifier(n_jobs=-1, random_state=10, min_samples_leaf=1, max_features=0.25)

rf.fit(df_train.drop(columns='Category'), df_train['Category'])

y_train_pred = rf.predict(df_train.drop(columns='Category'))
y_val_pred = rf.predict(df_val.drop(columns='Category'))

print('Train accuracy score: '
      , accuracy_score(df_train['Category'], y_train_pred)
)
print('Val accuracy score: '
      , accuracy_score(df_val['Category'], y_val_pred)
)

rf = RandomForestClassifier(n_jobs=-1, random_state=10, min_samples_leaf=1, max_features=0.75)

rf.fit(df_train.drop(columns='Category'), df_train['Category'])

y_train_pred = rf.predict(df_train.drop(columns='Category'))
y_val_pred = rf.predict(df_val.drop(columns='Category'))

print('Train accuracy score: '
      , accuracy_score(df_train['Category'], y_train_pred)
)
print('Val accuracy score: '
      , accuracy_score(df_val['Category'], y_val_pred)
)

dict(zip(rf.feature_names_in_, rf.feature_importances_))

# Test

df_test = pd.read_csv('data/processed/test.csv')
df_test

with open('models/enc.pickle', 'rb') as file:  
    enc = pickle.load(file) 

with open('models/fillna.pickle', 'rb') as file:  
    fillna = pickle.load(file) 

with open('models/tf_idf.pickle', 'rb') as file:  
    tf_idf = pickle.load(file) 

for col in fillna.keys():
    df_test.loc[df_test[col].isna(), col] = fillna[col]

df_test

enc

for c in enc.columns:
    enc.encode(df_test, c)
    
df_test.head()

tf_idf_matr_test = tf_idf.transform(df_test['Description'])

tf_idf_matr_test

df_tf_idf_test = pd.DataFrame(tf_idf_matr_test.toarray(), columns=tf_idf.get_feature_names_out())
df_tf_idf_test

df_concat_test = pd.concat([df_test, df_tf_idf_test], axis=1)
df_concat_test

df_concat_test.drop(columns=['Description'], inplace=True)

y_test_pred = rf.predict(df_concat_test.drop(columns='Category'))

print('Test accuracy score: '
      , accuracy_score(df_concat_test['Category'], y_test_pred)
)

# Save

with open('models/rf.pickle', 'wb') as handle:
    pickle.dump(rf, handle)
